made it look *pretty*

+++ bot.js
@@ -24,12 +24,17 @@ client.on('messageCreate', (message) => {
         return;
     }
 
-
-
     if (text.split(" ").length == 1) {
         if ((message.member) == '267680402594988033') {
             return
+            //whitelists akwav
         }
+
+        if ((message.channel) == client.channels.cache.get('920400419400863774')) {
+            return
+            //whitelists channel
+        }
+
         if (new Date() - messageTimes[message.author.id] < 10000) {
             message.delete()
         }
@@ -41,8 +46,11 @@ client.on('messageCreate', (message) => {
 
         if ((message.member) == '267680402594988033') {
             return
+            //whitelists akwav
         }
-        else return message.delete()
+        else {
+            return message.delete()
+        }
     }
 
     if (message.channel.id === process.env.CHANNEL_ID_SUPPORT) {
